andon/requirements.txt
streamlit==1.40.0
paho-mqtt==2.1.0
joblib==1.4.2
scikit-learn==1.5.2
lightgbm==4.5.0
pandas==2.2.3
